LayoutTests/fast/dom/SelectorAPI/caseID-strict.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
    <p id="description"></p>
    <div id="console"></div>
    <div style="display: none">
    <span id="lower1">lower 1</span><br>
    <span id="lower2">lower 2</span><br>
    <span id="UPPER1">UPPER 1</span><br>
    <span id="UPPER2">UPPER 2</span><br>
    </div>
    <script>
        shouldBe("document.querySelector('#lower1').textContent", "'lower 1'");
        shouldBeNull("document.querySelector('#LOWER2')");
        shouldBe("document.querySelector('#UPPER1').textContent", "'UPPER 1'");
        shouldBeNull("document.querySelector('#upper2')");

        shouldBeTrue("document.getElementById('lower1').webkitMatchesSelector('#lower1')");
        shouldBeFalse("document.getElementById('lower2').webkitMatchesSelector('#LOWER2')");
        shouldBeTrue("document.getElementById('UPPER1').webkitMatchesSelector('#UPPER1')");
        shouldBeFalse("document.getElementById('UPPER2').webkitMatchesSelector('#upper2')");

        var successfullyParsed = true;
    </script>
    <script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
