LayoutTests/fast/dom/SelectorAPI/caseID-strict-expected.txt

PASS document.querySelector('#lower1').textContent is 'lower 1'
PASS document.querySelector('#LOWER2') is null
PASS document.querySelector('#UPPER1').textContent is 'UPPER 1'
PASS document.querySelector('#upper2') is null
PASS document.getElementById('lower1').webkitMatchesSelector('#lower1') is true
PASS document.getElementById('lower2').webkitMatchesSelector('#LOWER2') is false
PASS document.getElementById('UPPER1').webkitMatchesSelector('#UPPER1') is true
PASS document.getElementById('UPPER2').webkitMatchesSelector('#upper2') is false
PASS successfullyParsed is true

TEST COMPLETE
